# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyloWare API",
    "description": "Llama Stack-native multi-agent video production platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/media/video/{video_id}": {
      "head": {
        "tags": [
          "media"
        ],
        "summary": "Head Video",
        "description": "HEAD request for video - allows URL validation without downloading.",
        "operationId": "head_video_v1_media_video__video_id__head",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get Video",
        "description": "Proxy Remotion rendered videos through the tunnel.\n\nThis allows external services (like Upload-Post) to access\nlocally rendered videos via the public tunnel URL.",
        "operationId": "get_video_v1_media_video__video_id__get",
        "parameters": [
          {
            "name": "video_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Video Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/media/transcoded/{filename}": {
      "head": {
        "tags": [
          "media"
        ],
        "summary": "Head Transcoded Video",
        "description": "HEAD request for transcoded OpenAI Sora video.",
        "operationId": "head_transcoded_video_v1_media_transcoded__filename__head",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get Transcoded Video",
        "description": "Serve transcoded OpenAI Sora videos.\n\nOpenAI Sora videos are transcoded to H.264/AAC on webhook receipt\nto ensure compatibility with Remotion's Chromium renderer.",
        "operationId": "get_transcoded_video_v1_media_transcoded__filename__get",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/media/sora/{task_id}.mp4": {
      "head": {
        "tags": [
          "media"
        ],
        "summary": "Head Fake Sora Video",
        "description": "HEAD request for fake Sora result URLs.\n\nIn `SORA_PROVIDER=fake`, the webhook payload includes `resultUrls` pointing\nat this endpoint to mimic Sora hosting a downloadable MP4.",
        "operationId": "head_fake_sora_video_v1_media_sora__task_id__mp4_head",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "media"
        ],
        "summary": "Get Fake Sora Video",
        "description": "Serve fake Sora MP4s for contract-exact local runs.",
        "operationId": "get_fake_sora_video_v1_media_sora__task_id__mp4_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}